requirements.txt
alembic==1.5.8
APScheduler==3.6.3
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
configparser==5.0.2
cryptography==3.4.7
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.2
Flask-WTF==0.14.3
greenlet==1.0.0
itsdangerous==1.1.0
jaraco.context==4.0.0
jaraco.docker==2.0
jaraco.functools==3.2.1
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
more-itertools==8.7.0
passlib==1.7.4
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
python-telegram-bot==13.4.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.2
tornado==6.1
tzlocal==2.1
Werkzeug==1.0.1
WTForms==2.3.3
